Try to get stat working in tests

+++ src/__tests__/decode.test.ts
@@ -11,8 +11,32 @@ const fs = {
 
 const log = false ? console.log : (a: unknown) => {};
 
-export function within(val: number, low: number, high: number): boolean {
+function within(val: number, low: number, high: number): boolean {
   return val >= low && val <= high;
+}
+
+function waitForFile(fileName: string): ofs.Stats | undefined {
+  let time = new Date().getSeconds();
+  do {
+    try {
+      const stat = fs.statSync(fileName);
+      return stat;
+    } catch (e) {}
+  } while (new Date().getSeconds() - time < 3);
+  return;
+}
+
+async function waitForFileAsync(
+  fileName: string,
+): Promise<ofs.Stats | undefined> {
+  let time = new Date().getSeconds();
+  do {
+    try {
+      const stat = await fs.statAsync(fileName);
+      return stat;
+    } catch (e) {}
+  } while (new Date().getSeconds() - time < 3);
+  return;
 }
 
 const cleanup = () => {
@@ -32,9 +56,10 @@ test('Async m4a to wav (using faad)', async () => {
     'src/__tests__/01-quiet.m4a',
     'src/__tests__/test-output1.wav',
   );
-  await Sleep(1000);
   expect(dec).toBe(true);
-  const stat = await fs.statAsync('src/__tests__/test-output1.wav');
+  const stat = await waitForFileAsync('src/__tests__/test-output1.wav');
+  expect(stat).toBeTruthy();
+  if (!stat) return;
   expect(within(stat.size, 88700, 90200)).toBeTruthy(); // Not great, but it works for now
   log(dec);
 });
@@ -43,9 +68,10 @@ test('Async mp3 to wav (using lame)', async () => {
     'src/__tests__/01-quiet.mp3',
     'src/__tests__/test-output2.wav',
   );
-  await Sleep(1000);
   expect(dec).toBe(true);
-  const stat = await fs.statAsync('src/__tests__/test-output2.wav');
+  const stat = await waitForFileAsync('src/__tests__/test-output2.wav');
+  expect(stat).toBeTruthy();
+  if (!stat) return;
   expect(stat.size).toBe(88788); // Not great, but it works for now
   log(dec);
 });
@@ -55,8 +81,9 @@ test('Async flac to wav', async () => {
     'src/__tests__/test-output3.wav',
   );
   expect(dec).toBe(true);
-  await Sleep(1000);
-  const stat = await fs.statAsync('src/__tests__/test-output3.wav');
+  const stat = await waitForFileAsync('src/__tests__/test-output3.wav');
+  expect(stat).toBeTruthy();
+  if (!stat) return;
   expect(within(stat.size, 133150, 133200)).toBeTruthy(); // Not great, but it works for now
   log(dec);
 });
@@ -66,8 +93,9 @@ test('Async wma to wav', async () => {
     'src/__tests__/test-output4.wav',
   );
   expect(dec).toBe(true);
-  await Sleep(1000);
-  const stat = await fs.statAsync('src/__tests__/test-output4.wav');
+  const stat = await waitForFileAsync('src/__tests__/test-output4.wav');
+  expect(stat).toBeTruthy();
+  if (!stat) return;
   expect(stat.size).toBe(90190); // Not great, but it works for now
   log(dec);
 });
@@ -78,7 +106,9 @@ test('Simple m4a to wav (using faad)', () => {
     'src/__tests__/test-output5.wav',
   );
   expect(dec).toBe(true);
-  const stat = fs.statSync('src/__tests__/test-output5.wav');
+  const stat = waitForFile('src/__tests__/test-output5.wav');
+  expect(stat).toBeTruthy();
+  if (!stat) return;
   expect(within(stat.size, 88700, 90200)).toBeTruthy(); // Not great, but it works for now
   log(dec);
 });
@@ -88,7 +118,9 @@ test('Simple mp3 to wav (using lame)', () => {
     'src/__tests__/test-output6.wav',
   );
   expect(dec).toBe(true);
-  const stat = fs.statSync('src/__tests__/test-output6.wav');
+  const stat = waitForFile('src/__tests__/test-output6.wav');
+  expect(stat).toBeTruthy();
+  if (!stat) return;
   expect(stat.size).toBe(88788); // Not great, but it works for now
   log(dec);
 });
@@ -98,7 +130,9 @@ test('Simple flac to wav', () => {
     'src/__tests__/test-output7.wav',
   );
   expect(dec).toBe(true);
-  const stat = fs.statSync('src/__tests__/test-output7.wav');
+  const stat = waitForFile('src/__tests__/test-output7.wav');
+  expect(stat).toBeTruthy();
+  if (!stat) return;
   expect(within(stat.size, 133150, 133200)).toBeTruthy(); // Not great, but it works for now
   log(dec);
 });
@@ -108,7 +142,9 @@ test('Simple wma to wav', () => {
     'src/__tests__/test-output8.wav',
   );
   expect(dec).toBe(true);
-  const stat = fs.statSync('src/__tests__/test-output8.wav');
+  const stat = waitForFile('src/__tests__/test-output8.wav');
+  expect(stat).toBeTruthy();
+  if (!stat) return;
   expect(stat.size).toBe(90190); // Not great, but it works for now
   log(dec);
 });
@@ -120,7 +156,9 @@ for (let type of types) {
     expect(dec).toBeDefined();
     if (dec) {
       await Sleep(1000);
-      const stat = await fs.statAsync(dec);
+      const stat = await waitForFileAsync(dec);
+      expect(stat).toBeTruthy();
+      if (!stat) return;
       expect(within(stat.size, 88000, 135000)).toBeTruthy();
       await fs.unlinkAsync(dec);
     }
@@ -129,7 +167,9 @@ for (let type of types) {
     const dec = Decode.MakeWave(`src/__tests__/01-quiet.${type}`);
     expect(dec).toBeDefined();
     if (dec) {
-      const stat = fs.statSync(dec);
+      const stat = waitForFile(dec);
+      expect(stat).toBeTruthy();
+      if (!stat) return;
       expect(within(stat.size, 88000, 135000)).toBeTruthy();
       fs.unlinkSync(dec);
     }
